Bit of test refactoring
